Partially rewrite NTSC Palette generator
Get rid of the monstrosity that was the getPhase() function, and
replace it with proper code for calculating YIQ based on information on
the NESdev wiki.

Note that this still uses the old YIQ coefficients and does not perform
gamma correction, so the "standard" palette should look the same as
before; however, the Emphasis palettes are now noticeably different
(and arguably a lot closer to what they should be).